+++ src/providers/AppProvider.tsx
@@ -6,6 +6,7 @@ import {
   RefObject,
 } from 'react';
 import {
+  COOKIES_TYPE,
   LANG_TYPE,
   LAYOUT_TYPE,
   NAVIGATION_PATHS,
@@ -15,6 +16,8 @@ import { useLocation, useNavigate, useSearchParams } from 'react-router-dom';
 import { IContextProviderParams } from '../utils/interfaces/contexts';
 import tailwindConfig from '../../tailwind.config';
 import i18next from '../utils/i18n/i18next';
+import { useCookies } from 'react-cookie';
+import useCookiesContext from '../hooks/contexts/useCookiesContext';
 
 export interface IAppContext {
   theme: THEME_TYPE;
@@ -60,10 +63,6 @@ export interface IAppContext {
 }
 
 export const AppContext = createContext<IAppContext | null>(null);
-// LOCAL SOTRAGE KEY
-const THEME_KEY = 'elvira-theme';
-const LANG_KEY = 'elvira-lang';
-const LAYOUT_KEY = 'elvira-layout';
 
 // IMAGES / LOGOS
 const logoDark = `/assets/${
@@ -86,23 +85,18 @@ const stuLogoLight = `/assets/${
 }/stu/logo-light.png`;
 
 const AppProvider = ({ children }: IContextProviderParams) => {
-  // Set initail value from theme and lang
-  const getInitialTheme = () => {
-    const theme = localStorage.getItem(THEME_KEY);
-    return theme ? JSON.parse(theme) : THEME_TYPE.light;
-  };
-  const getInitialLang = () => {
-    const lang = localStorage.getItem(LANG_KEY);
-    return lang ? JSON.parse(lang) : LANG_TYPE.sk;
-  };
-  const getInitialLayout = () => {
-    const layout = localStorage.getItem(LAYOUT_KEY);
-    return layout ? JSON.parse(layout) : LAYOUT_TYPE.box;
-  };
   const { colors } = tailwindConfig.theme?.extend!;
-  const [theme, setTheme] = useState<THEME_TYPE>(getInitialTheme);
-  const [lang, setLang] = useState<LANG_TYPE>(getInitialLang);
-  const [layout, setLayout] = useState<LAYOUT_TYPE>(getInitialLayout);
+  const { cookies, setCookie } = useCookiesContext();
+
+  const [theme, setTheme] = useState<THEME_TYPE>(
+    cookies[COOKIES_TYPE.THEME_KEY] ?? THEME_TYPE.light
+  );
+  const [lang, setLang] = useState<LANG_TYPE>(
+    cookies[COOKIES_TYPE.LANG_KEY] ?? LANG_TYPE.sk
+  );
+  const [layout, setLayout] = useState<LAYOUT_TYPE>(
+    cookies[COOKIES_TYPE.LAYOUT_KEY] ?? LAYOUT_TYPE.box
+  );
   const [showNavbar, setShowNavbar] = useState<boolean>(false);
   const [editingEntryTitle, setEditingEntryTitle] = useState<string>('');
   const [showSearchBar, setShowSearchBar] = useState<boolean>(false);
@@ -117,19 +111,19 @@ const AppProvider = ({ children }: IContextProviderParams) => {
   // Update theme and localstorage
   const updateTheme = (theme: THEME_TYPE) => {
     setTheme(theme);
-    localStorage.setItem(THEME_KEY, JSON.stringify(theme));
+    setCookie(COOKIES_TYPE.THEME_KEY, theme, { maxAge: 60 * 60 * 24 * 365 }); // year
   };
 
   // Update lang and localstorage
   const updateLang = (lang: LANG_TYPE) => {
     setLang(lang);
-    localStorage.setItem(LANG_KEY, JSON.stringify(lang));
+    setCookie(COOKIES_TYPE.LANG_KEY, lang, { maxAge: 60 * 60 * 24 * 365 }); // year
   };
 
   // Update layout and localstorage
   const updateLayout = (layout: LAYOUT_TYPE) => {
     setLayout(layout);
-    localStorage.setItem(LAYOUT_KEY, JSON.stringify(layout));
+    setCookie(COOKIES_TYPE.LAYOUT_KEY, layout, { maxAge: 60 * 60 * 24 * 365 }); // year
   };
 
   // Special navigation stands for navigation that can open new window tab with holding ctr/cmd